UI Phase 2 complete: Dogs, People, Photos, Maps, Messaging, Theme polish

+++ lib/pages/vehicle_report_page.dart
@@ -219,9 +219,9 @@ class _VehicleReportPageState extends State<VehicleReportPage> {
 
     final logs = await loadLogs();
 
-    int totalKm = 0;
-    int businessKm = 0;
-    int privateKm = 0;
+    int totalKm = 0.toInt();
+    int businessKm = 0.toInt();
+    int privateKm = 0.toInt();
 
     final pdf = pw.Document();
